finished Async tx & rx in usart and it need a bet of optimization

+++ NVIC/NVIC_INTERFACE.h
@@ -9,6 +9,8 @@ Github      : https://github.com/KerolosMaged
 #define NVIC_INTERFACE_H
 
 #include "../LIB/STD_TYPES_MATH.h"
+#include "NVIC_CONFGR.h"
+#include "NVIC_PRIVATE.h"
 
 void NVIC_EnableIRQ(IRQn_Type IRQn);
 
@@ -18,11 +20,11 @@ void NVIC_SetPendingIRQ(IRQn_Type IRQn);
 
 void NVIC_ClearPendingIRQ(IRQn_Type IRQn);
 
-uint32_t NVIC_GetPendingIRQ(IRQn_Type IRQn);
+uint8_t NVIC_GetPendingIRQ(IRQn_Type IRQn);
 
 void NVIC_SetPriority(IRQn_Type IRQn, uint32_t priority);
 
-uint32_t NVIC_GetPriority(IRQn_Type IRQn);
+uint8_t NVIC_GetPriority(IRQn_Type IRQn);
